feat: switch to default no movement approach
switches to a "no news is good news" approach, where the default
value is stationary. This is to allow for continued use of sparse
optical flow (rather than dense), but ensures that only the moving
corner points are considered as part of the average, rather than
points not on the moving object dragging down the average.